add constraint in all CRUD'S

+++ controllers/userController.js
@@ -20,6 +20,12 @@ exports.userRegister = async (req, res) => {
 // méthode pour se connecter
 exports.userLogin = async (req, res) => {
     try {
+        // Vérifier si l'emailest unique
+        const existingEmail = await User.findOne({ email: req.body.email });
+        if (existingEmail) {
+            return res.status(400).json({ message: 'L email doit être unique.' });
+        }
+
         const user = await User.findOne({email: req.body.email});
         if(!user) {
             res.status(500).json({message: 'Utilisateur non trouvé'});
@@ -51,6 +57,12 @@ exports.userLogin = async (req, res) => {
 // méthode pour supprimer un utilisateur
 exports.deleteUser = async (req, res) => {
     try {
+        const user = await User.findById(req.params.user_id);
+
+        if (!user) {
+            return res.status(404).json({ message: 'Utilisateur non trouvé' });
+        }
+
         await User.findByIdAndDelete(req.params.user_id);
         res.status(200).json({message: 'Utilisateur supprimé'});
     } catch (error) {
@@ -62,6 +74,12 @@ exports.deleteUser = async (req, res) => {
 // méthode pour modifier partiellement un utilisateur
 exports.putUser = async (req, res) => {
     try {
+        // Vérifier si l'emailest unique
+        const existingEmail = await User.findOne({ email: req.body.email });
+        if (existingEmail) {
+            return res.status(400).json({ message: 'L email doit être unique.' });
+        }
+
         const user = await User.findByIdAndUpdate(req.params.user_id, req.body, {new: true});
         res.status(200).json(user);
     } catch (error) {
@@ -72,6 +90,12 @@ exports.putUser = async (req, res) => {
 // méthode pour modifier un utilisateur
 exports.patchUser = async (req, res) => {
     try {
+        // Vérifier si l'emailest unique
+        const existingEmail = await User.findOne({ email: req.body.email });
+        if (existingEmail) {
+            return res.status(400).json({ message: 'L email doit être unique.' });
+        }
+        
         const user = await User.findByIdAndUpdate(req.params.user_id, req.body, {new: true});
         res.status(200).json(user);
     } catch (error) {
@@ -94,7 +118,7 @@ exports.getUser = async (req, res) => {
 // methode pour les infos d'un groupe
 exports.getInfoGroup = async (req, res) => {
     try {
-        const allInfo = await Group.find({user_id: req.params.user_id});
+        const allInfo = await Group.find({group_id: req.params.group_id});
         if (allInfo.length == 0) {
             res.status(500).json({message: 'Aucun groupe trouvé.'});
         } else {
@@ -109,10 +133,17 @@ exports.getInfoGroup = async (req, res) => {
 exports.createGroup = async (req, res) => {
     try {
         await User.findById(req.params.user_id);
+
+        // Vérifier si le nom du groupe est unique
+        const existingGroup = await Group.findOne({ name: req.body.name });
+        if (existingGroup) {
+            return res.status(400).json({ message: 'Le nom du groupe doit être unique.' });
+        }
+
         const newGroup = new Group({...req.body, user_id: req.params.user_id});
         try {
             const group = await newGroup.save();
-            res.status(201).json(group);
+            res.status(201).json({ message: `Groupe créé ! id : ${group.id}` });  
         } catch (error) {
             res.status(500).json({message: 'Erreur serveur'});
         }
@@ -124,18 +155,30 @@ exports.createGroup = async (req, res) => {
 // méthode pour supprimer un groupe
 exports.deleteGroup = async (req, res) => {
     try {
-        await Group.findByIdAndDelete(req.params.user_id);
-        res.status(200).json({message: 'Groupe supprimé'});
+        const group = await Group.findById(req.params.group_id);
+
+        if (!group) {
+            return res.status(404).json({ message: 'Groupe non trouvé' });
+        }
+
+        await Group.findByIdAndDelete(req.params.group_id);
+        res.status(200).json({ message: 'Groupe supprimé' });
     } catch (error) {
-        console.log(error);
-        res.status(500).json({message: 'Erreur serveur'});
+        console.error(error);
+        res.status(500).json({ message: 'Erreur serveur' });
     }
 };
 
 // méthode pour modifier partiellement un groupe
 exports.putGroup = async (req, res) => {
     try {
-        const group = await Group.findByIdAndUpdate(req.params.user_id, req.body, {new: true});
+        // Vérifier si le nom du groupe est unique
+        const existingGroup = await Group.findOne({ name: req.body.name });
+        if (existingGroup) {
+            return res.status(400).json({ message: 'Le nom du groupe doit être unique.' });
+        }
+
+        const group = await Group.findByIdAndUpdate(req.params.group_id, req.body, {new: true});
         res.status(200).json(group);
     } catch (error) {
         res.status(500).json({message: 'Erreur serveur'});
@@ -145,7 +188,13 @@ exports.putGroup = async (req, res) => {
 // méthode pour modifier un groupe
 exports.patchGroup = async (req, res) => {
     try {
-        const group = await Group.findByIdAndUpdate(req.params.user_id, req.body, {new: true});
+        // Vérifier si le nom du groupe est unique
+        const existingGroup = await Group.findOne({ name: req.body.name });
+        if (existingGroup) {
+            return res.status(400).json({ message: 'Le nom du groupe doit être unique.' });
+        }
+
+        const group = await Group.findByIdAndUpdate(req.params.group_id, req.body, {new: true});
         res.status(200).json(group);
     } catch (error) {
         res.status(500).json({message: 'Erreur serveur'});
